requirements.txt
python-telegram-bot==13.15
gspread==5.12.0
google-auth==2.28.1
requests